Refactor server to simplify code
