+++ server-deps.ts
@@ -1,0 +1,1 @@
+export { serve } from 'https://deno.land/std@0.114.0/http/server.ts';
